requirements.txt
# ── LLM + Agent ───────────────────────────────────────────────────────────────
langchain==0.2.16
langchain-openai==0.1.23
langchain-community==0.2.16
langgraph==0.2.28

# ── Vector DB ─────────────────────────────────────────────────────────────────
chromadb==0.5.5

# ── Embeddings (free local fallback) ──────────────────────────────────────────
sentence-transformers==3.0.1

# ── Database ──────────────────────────────────────────────────────────────────
# sqlite3 is built into Python — no install needed!

# ── Backend API ───────────────────────────────────────────────────────────────
fastapi==0.112.2
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-dotenv==1.0.1

# ── Frontend ──────────────────────────────────────────────────────────────────
streamlit==1.37.1
pandas==2.2.2
requests==2.32.3
